Removed unstable code for large sets
The previous code allowed a one trillion cell pivot but the behaviour
became unstable on very small sets.  The problem is only down to browser
restrictions on div size, the most restrictive being IE.  The solution
is to factor the scrolling so that a move of 1 pixel on the scroll bar =
10000 pixels (for example) on the virtual grid.  This works but caused
other problems and is only really as cosmetic fix as it becomes
impossible to get to rows between the scroll positions.  So on a 10
million row table a single tick jumps 10000 rows and all the intervening
rows are unaccessable.  This commit also includes some new adapters.