Added Event Handling Feature
New external function on()  is introduced that gathers events that need
to be registered.

LayoutDOM invokes redirectViewportEvents to register a callback that
redirects the events to the target the event is meant for.

RenderRegion invokes addEventHandler to register the eventHandlers that
are gathered through on() function.

data() is modified to refresh conditionally so that on() can be called
to register events and then refresh() can be called to render.